backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.48
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
fastembed==0.3.5
yfinance==0.2.54
pypdf>=4.0.0